chromium/third_party/blink/web_tests/external/wpt/css/css-transforms/animation/transform-interpolation-computed-value-expected.txt

This is a testharness.js-based test.
Found 87 tests; 36 PASS, 51 FAIL, 0 TIMEOUT, 0 NOTRUN.
FAIL Interpolation between translateX(0px) and translateX(50px) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "translateX(25px)" but got "translate(25px, 0px)"
FAIL Interpolation between translateX(0px) and translateX(50px) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "translateX(25px)" but got "translate(25px, 0px)"
FAIL Interpolation between translateX(0px) and translateX(50px) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "translateX(25px)" but got "matrix(1, 0, 0, 1, 25, 0)"
FAIL Interpolation between translateX(0%) and translateX(50%) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "translateX(25%)" but got "translate(25%, 0px)"
FAIL Interpolation between translateX(0%) and translateX(50%) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "translateX(25%)" but got "translate(25%, 0px)"
FAIL Interpolation between translateX(0%) and translateX(50%) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "translateX(25%)" but got "matrix(1, 0, 0, 1, 196, 0)"
PASS Interpolation between translateY(0%) and translateX(50%) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between translateY(0%) and translateX(50%) gives the correct computed value halfway according to computedStyleMap with zoom active.
FAIL Interpolation between translateY(0%) and translateX(50%) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "translate(25%, 0px)" but got "matrix(1, 0, 0, 1, 196, 0)"
PASS Interpolation between translateX(50px) and translateY(50px) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between translateX(50px) and translateY(50px) gives the correct computed value halfway according to computedStyleMap with zoom active.
FAIL Interpolation between translateX(50px) and translateY(50px) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "translate(25px, 25px)" but got "matrix(1, 0, 0, 1, 25, 25)"
PASS Interpolation between translateX(50px) and translateZ(50px) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between translateX(50px) and translateZ(50px) gives the correct computed value halfway according to computedStyleMap with zoom active.
FAIL Interpolation between translateX(50px) and translateZ(50px) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "translate3d(25px, 0px, 25px)" but got "matrix3d(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 25, 0, 25, 1)"
PASS Interpolation between translateZ(50px) and translateX(50px) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between translateZ(50px) and translateX(50px) gives the correct computed value halfway according to computedStyleMap with zoom active.
FAIL Interpolation between translateZ(50px) and translateX(50px) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "translate3d(25px, 0px, 25px)" but got "matrix3d(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 25, 0, 25, 1)"
FAIL Interpolation between translateZ(-50px) and translateZ(50px) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "translateZ(0px)" but got "translate3d(0px, 0px, 0px)"
FAIL Interpolation between translateZ(-50px) and translateZ(50px) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "translateZ(0px)" but got "translate3d(0px, 0px, 0px)"
FAIL Interpolation between translateZ(-50px) and translateZ(50px) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "translateZ(0px)" but got "matrix(1, 0, 0, 1, 0, 0)"
PASS Interpolation between translate3d(0,0,-50px) and translateZ(50px) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between translate3d(0,0,-50px) and translateZ(50px) gives the correct computed value halfway according to computedStyleMap with zoom active.
FAIL Interpolation between translate3d(0,0,-50px) and translateZ(50px) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "translate3d(0px, 0px, 0px)" but got "matrix(1, 0, 0, 1, 0, 0)"
PASS Interpolation between rotate(30deg) and rotate(90deg) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between rotate(30deg) and rotate(90deg) gives the correct computed value halfway according to computedStyleMap with zoom active.
FAIL Interpolation between rotate(30deg) and rotate(90deg) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "rotate(60deg)" but got "matrix(0.5, 0.866025, -0.866025, 0.5, 0, 0)"
FAIL Interpolation between rotateZ(30deg) and rotateZ(90deg) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "rotateZ(60deg)" but got "rotate(60deg)"
FAIL Interpolation between rotateZ(30deg) and rotateZ(90deg) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "rotateZ(60deg)" but got "rotate(60deg)"
FAIL Interpolation between rotateZ(30deg) and rotateZ(90deg) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "rotateZ(60deg)" but got "matrix(0.5, 0.866025, -0.866025, 0.5, 0, 0)"
FAIL Interpolation between rotate(0deg) and rotateZ(90deg) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "rotate3d(0, 0, 1, 45deg)" but got "rotate(45deg)"
FAIL Interpolation between rotate(0deg) and rotateZ(90deg) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "rotate3d(0, 0, 1, 45deg)" but got "rotate(45deg)"
FAIL Interpolation between rotate(0deg) and rotateZ(90deg) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "rotate3d(0, 0, 1, 45deg)" but got "matrix(0.707107, 0.707107, -0.707107, 0.707107, 0, 0)"
FAIL Interpolation between rotateX(0deg) and rotateX(90deg) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "rotateX(45deg)" but got "rotate3d(1, 0, 0, 45deg)"
FAIL Interpolation between rotateX(0deg) and rotateX(90deg) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "rotateX(45deg)" but got "rotate3d(1, 0, 0, 45deg)"
FAIL Interpolation between rotateX(0deg) and rotateX(90deg) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "rotateX(45deg)" but got "matrix3d(1, 0, 0, 0, 0, 0.707107, 0.707107, 0, 0, -0.707107, 0.707107, 0, 0, 0, 0, 1)"
PASS Interpolation between rotate(0deg) and rotateX(90deg) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between rotate(0deg) and rotateX(90deg) gives the correct computed value halfway according to computedStyleMap with zoom active.
FAIL Interpolation between rotate(0deg) and rotateX(90deg) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "rotate3d(1, 0, 0, 45deg)" but got "matrix3d(1, 0, 0, 0, 0, 0.707107, 0.707107, 0, 0, -0.707107, 0.707107, 0, 0, 0, 0, 1)"
FAIL Interpolation between scale(1) and scale(2) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "scale(1.5)" but got "scale(1.5, 1.5)"
FAIL Interpolation between scale(1) and scale(2) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "scale(1.5)" but got "scale(1.5, 1.5)"
FAIL Interpolation between scale(1) and scale(2) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "scale(1.5)" but got "matrix(1.5, 0, 0, 1.5, 0, 0)"
FAIL Interpolation between scaleX(1) and scaleX(2) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "scaleX(1.5)" but got "scale(1.5, 1)"
FAIL Interpolation between scaleX(1) and scaleX(2) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "scaleX(1.5)" but got "scale(1.5, 1)"
FAIL Interpolation between scaleX(1) and scaleX(2) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "scaleX(1.5)" but got "matrix(1.5, 0, 0, 1, 0, 0)"
FAIL Interpolation between scaleY(1) and scaleY(2) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "scaleY(1.5)" but got "scale(1, 1.5)"
FAIL Interpolation between scaleY(1) and scaleY(2) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "scaleY(1.5)" but got "scale(1, 1.5)"
FAIL Interpolation between scaleY(1) and scaleY(2) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "scaleY(1.5)" but got "matrix(1, 0, 0, 1.5, 0, 0)"
FAIL Interpolation between scaleZ(1) and scaleZ(2) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "scaleZ(1.5)" but got "scale3d(1, 1, 1.5)"
FAIL Interpolation between scaleZ(1) and scaleZ(2) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "scaleZ(1.5)" but got "scale3d(1, 1, 1.5)"
FAIL Interpolation between scaleZ(1) and scaleZ(2) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "scaleZ(1.5)" but got "matrix3d(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1.5, 0, 0, 0, 0, 1)"
FAIL Interpolation between scaleX(2) and scaleY(2) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "scale(1.5)" but got "scale(1.5, 1.5)"
FAIL Interpolation between scaleX(2) and scaleY(2) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "scale(1.5)" but got "scale(1.5, 1.5)"
FAIL Interpolation between scaleX(2) and scaleY(2) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "scale(1.5)" but got "matrix(1.5, 0, 0, 1.5, 0, 0)"
PASS Interpolation between scaleX(2) and scaleY(3) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between scaleX(2) and scaleY(3) gives the correct computed value halfway according to computedStyleMap with zoom active.
FAIL Interpolation between scaleX(2) and scaleY(3) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "scale(1.5, 2)" but got "matrix(1.5, 0, 0, 2, 0, 0)"
PASS Interpolation between scaleZ(1) and scale(2) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between scaleZ(1) and scale(2) gives the correct computed value halfway according to computedStyleMap with zoom active.
FAIL Interpolation between scaleZ(1) and scale(2) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "scale3d(1.5, 1.5, 1)" but got "matrix(1.5, 0, 0, 1.5, 0, 0)"
PASS Interpolation between scale(1, 2) and scale(3, 4) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between scale(1, 2) and scale(3, 4) gives the correct computed value halfway according to computedStyleMap with zoom active.
FAIL Interpolation between scale(1, 2) and scale(3, 4) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "scale(2, 3)" but got "matrix(2, 0, 0, 3, 0, 0)"
PASS Interpolation between scale3d(1, 2, 3) and scale3d(4, 5, 6) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between scale3d(1, 2, 3) and scale3d(4, 5, 6) gives the correct computed value halfway according to computedStyleMap with zoom active.
FAIL Interpolation between scale3d(1, 2, 3) and scale3d(4, 5, 6) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "scale3d(2.5, 3.5, 4.5)" but got "matrix3d(2.5, 0, 0, 0, 0, 3.5, 0, 0, 0, 0, 4.5, 0, 0, 0, 0, 1)"
PASS Interpolation between scale3d(1, 2, 3) and scale(4, 5) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between scale3d(1, 2, 3) and scale(4, 5) gives the correct computed value halfway according to computedStyleMap with zoom active.
FAIL Interpolation between scale3d(1, 2, 3) and scale(4, 5) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "scale3d(2.5, 3.5, 2)" but got "matrix3d(2.5, 0, 0, 0, 0, 3.5, 0, 0, 0, 0, 2, 0, 0, 0, 0, 1)"
PASS Interpolation between scale(1, 2) and scale3d(3, 4, 5) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between scale(1, 2) and scale3d(3, 4, 5) gives the correct computed value halfway according to computedStyleMap with zoom active.
FAIL Interpolation between scale(1, 2) and scale3d(3, 4, 5) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "scale3d(2, 3, 3)" but got "matrix3d(2, 0, 0, 0, 0, 3, 0, 0, 0, 0, 3, 0, 0, 0, 0, 1)"
PASS Interpolation between skewX(0deg) and skewX(60deg) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between skewX(0deg) and skewX(60deg) gives the correct computed value halfway according to computedStyleMap with zoom active.
FAIL Interpolation between skewX(0deg) and skewX(60deg) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "skewX(30deg)" but got "matrix(1, 0, 0.57735, 1, 0, 0)"
PASS Interpolation between skewX(0deg) and skewX(90deg) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between skewX(0deg) and skewX(90deg) gives the correct computed value halfway according to computedStyleMap with zoom active.
FAIL Interpolation between skewX(0deg) and skewX(90deg) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "skewX(45deg)" but got "matrix(1, 0, 1, 1, 0, 0)"
PASS Interpolation between skewX(0deg) and skewX(180deg) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between skewX(0deg) and skewX(180deg) gives the correct computed value halfway according to computedStyleMap with zoom active.
FAIL Interpolation between skewX(0deg) and skewX(180deg) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "skewX(90deg)" but got "matrix(1, 0, 1.63312e+16, 1, 0, 0)"
PASS Interpolation between skew(0deg, 0deg) and skew(60deg, 60deg) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between skew(0deg, 0deg) and skew(60deg, 60deg) gives the correct computed value halfway according to computedStyleMap with zoom active.
FAIL Interpolation between skew(0deg, 0deg) and skew(60deg, 60deg) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "skew(30deg, 30deg)" but got "matrix(1, 0.57735, 0.57735, 1, 0, 0)"
PASS Interpolation between skew(45deg, 0deg) and skew(0deg, 45deg) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between skew(45deg, 0deg) and skew(0deg, 45deg) gives the correct computed value halfway according to computedStyleMap with zoom active.
FAIL Interpolation between skew(45deg, 0deg) and skew(0deg, 45deg) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "skew(22.5deg, 22.5deg)" but got "matrix(1, 0.414214, 0.414214, 1, 0, 0)"
Harness: the test ran to completion.
